arc: ignore file type for decompression
Archives are either compressed or not compressed, which is controlled by
the ARCHIVE_RAW flag, not by the file extension of archived files.

This is supposed to be controlled by the bMESTYPE and bDATATYPE values
in AI5WIN.ini. libai5 does not read these values and depends on the
caller to specify the correct behavior.

This change is needed for Kakyuusei, which has a compressed MES archive
and an uncompressed DATA archive.