mes: initial AIWIN engine support
Support parsing/printing .mes files from 'Shuusaku' and 'Kawarazaki-ke
no Ichizoku', which use the 'AIWIN' engine (as opposed to 'AI5WIN').

The bytecode for this engine is similar to AI5WIN bytecode (for example,
most expression opcodes are identical) but differs in many details.
Statement opcodes are completely different, and most of what fell under
the 'System' namespace in AI5WIN has dedicated opcodes in AIWIN.

The same structures are used to represent parsed AIWIN and AI5WIN
bytecode, but with the opcodes interpreted differently based on which
engine is being targeted.